requirements.txt
pyxdg
requests
quebra_frases